+++ src/bitcoinrpc.cpp
@@ -6,6 +6,7 @@
 #include "main.h"
 #include "wallet.h"
 #include "db.h"
+#include "walletdb.h"
 #include "net.h"
 #include "init.h"
 #include "ui_interface.h"
