+++ src/makefile.linux-mingw
@@ -62,6 +62,7 @@ OBJS= \
     obj/script.o \
     obj/util.o \
     obj/wallet.o \
+    obj/walletdb.o \
     obj/noui.o
 
 all: bitcoind.exe
